% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove.outliers}
\alias{remove.outliers}
\title{Remove outliers from a vector}
\usage{
remove.outliers(x, multiplier = 1.5)
}
\arguments{
\item{x}{A numeric vector}

\item{multiplier}{step multiplier, defaults to 1.5}
}
\value{
A numeric vector without outliers & NA values
}
\description{
Removes the outliers from a numeric vector & ignores NA values.
}
\examples{
x <- c(4,5,5,8,8,9,5,2,5,6,7,8,4,5,99)
remove.outliers(x)
}

