% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_split.R
\name{train.test.split}
\alias{train.test.split}
\title{Train test split}
\usage{
train.test.split(data, test.size = 0, train.size = 0, shuffle = TRUE)
}
\arguments{
\item{data}{A dataframe to split into test and train}

\item{test.size}{Either an integer indicating the number of rows that should be in test, or a percentage (expressed as a number between 0 and 1) of the number of rows from data that should be in test}

\item{train.size}{Either an integer indicating the number of rows that should be in train, or a percentage (expressed as a number between 0 and 1) of the number of rows from data that should be in train}

\item{shuffle}{A boolean indicating if the dataframe should be shuffled, defaults to TRUE}
}
\value{
A named list where 'test' is the test dataframe and 'train' is the train dataframe
}
\description{
Split a dataframe into two dataframes, one for testing and one for training. Very similar to sklearn.model_selection.train_test_split
}
\examples{
data("mtcars")
train.test.split(mtcars, train.size = 0.7)
train.test.split(mtcars, test.size = 14)
}

